backend/requirements.txt
fastapi==0.116.1
httpx==0.28.1
motor==3.7.1
pydantic==2.11.7
PyJWT==2.10.1
PyJWT==2.10.1
python-dotenv==1.1.1
uvicorn==0.35.0
